Create README.md
This is a simple Harry Potter Quiz I have developed using HTML, CSS and JavaScript. The quiz is of 10 questions with questions displaying one at a time followed by four options out of which one is the correct answer. The correct answer is shown in green color and rest of them are shown in red color. For each correct answer, you will be given 1 mark and the score will be displayed at the top-right corner for every question. After all the questions are finished, the final score will be displayed and you can restart the quiz again for which the button is provided.